requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
flask-cors==4.0.0
